closes #9 : add IE support using method proposed by Manu Sharma in http://stackoverflow.com/a/27699027/103081
added dltest.html to main directory for manual download test similar to examples
1 and 2 in the README.md

tested -- working and generated .csv opens in excel-- in IE 11 on Win 7
tested -- working -- in Firefox and Chrome on Ubuntu 14.04 LTS

TODO: Qunit Tests 4.2,4.3 are failing on all platforms.  This is unrelated to
downloading, but are minor "should throw" parameter-checking tests in the code
to  support user-extensions.