This closes #11 Fix CSV global not initialized
This was a conflict with "use strict;" which, among other things, disallows the creation of new globals.
That's why "window.CSV = " was fine but "CSV = " was broken, as the former assigned to an existing global,
i.e. "window", whereas the latter created a global directly.

An earlier commit to enable usage from Meteor required defining the global directly so that Meteor's code wrangler
could decide whether to inject it as a dependency in other Meteor modules.