Cleanup CSV.extend() handling and disable testing of (null) or (null,null) parameters
In html5csv.js:  use a nested if to avoid calling jquery $.extend when parameters are null.

In qtestcsv.js:  do not use (null) and (null,null) as test cases expecting a throw.

It is unclear whether CSV.extend(null,null) should throw.  Here we decided to interpret it as a do nothing.